+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.2.0] - 2019-07-03
+### Added
+- -File parameter, to optionally specify an XML file instead of the built-in template
+- Better documentation in the form of comment-based help
+
 ## [0.1.0] - 2019-07-01
 ### Added
 - First push to Github